Plots brains in 3D

brainconn plots and returns a interactive plotly object of plotted brain connectivity matrix..

brainconn3D(
  atlas = NULL,
  conmat = NULL,
  all.nodes = FALSE,
  node.color = "network",
  node.size = 10,
  edge.width = 3,
  edge.color = "black",
  opacity = 0.28,
  d.factor = 1.05,
  show.legend = F,
  thr = NULL,
  uthr = NULL
)

Arguments

conmat

A adjacency matrix. Has ti be binary and undirected (i.e. symetric). see example_* data.

all.nodes

f TRUE, then all nodes will be shown be hemisphere without ticks. If FALSE, then only nodes with connecting edges will be shown.

node.color

A string that sets the node color. e.g. "blue". If set to "network", then nodes will be colored according to the network column of the atlas

node.size

A integer that determines the diameter of the nodes. Can also be a vector of integers with a length equal to the number of ROIs in the atlas

edge.width

Number to set the width of the edges.

edge.color

A string that sets the edge color. e.g. "blue".

opacity

Number between 0-1 to set the transparency of the mesh brain

d.factor

An int to change the distance between the nodes and brain surface.

show.legend

if TRUE, legend will be shown. If FALSE, then no legend will be shown

thr

a optional value to set a threshold on the conmat (e.g. edges with a weighted value lower than the one set here will not be shown)

uthr

a optional value to set a upper threshold on the conmat (e.g. edges with a weighted value higher than the one set here will not be shown)

Value

a plotly object

Author

Sidhant Chopra

Examples

if (FALSE) {
library(brainconn)
brainconn3D(atlas ="schaefer300_n7", conmat=example_unweighted_undirected, show.legend = F)
}